refactor: use svgr react imports
