+++ arsh-game/src/routes/Dashboard.tsx
@@ -1,8 +1,8 @@
 import { useRef, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { ReactComponent as MindSVG } from '@/assets/svg/mind.svg'
-import { ReactComponent as BodySVG } from '@/assets/svg/body.svg'
-import { ReactComponent as SpiritSVG } from '@/assets/svg/spirit.svg'
+import MindSVG from '@/assets/svg/mind.svg?react'
+import BodySVG from '@/assets/svg/body.svg?react'
+import SpiritSVG from '@/assets/svg/spirit.svg?react'
 import Hero from '@/components/Hero'
 import Panel from '@/components/Panel'
 import Aura from '@/components/Aura'
