+++ arsh-game/src/vite-env.d.ts
@@ -1,9 +1,3 @@
 /// <reference types="vite/client" />
+/// <reference types="vite-plugin-svgr/client" />
 
-import type { FunctionComponent, SVGProps } from 'react'
-
-declare module '*.svg' {
-  export const ReactComponent: FunctionComponent<SVGProps<SVGSVGElement>>
-  const src: string
-  export default src
-}
